Keep old versions of ThroughputSOSS and WebbKer in engine_legacy.py, so other code doesn't break.
